+++ src/display.hpp
@@ -7,7 +7,7 @@
 namespace gp {
 
 using KeyCallback = void(*)(void* userdata, int key, int action);
-constexpr const long kMaxKeyCallbacks = 1;
+constexpr const int kMaxKeyCallbacks = 1;
 
 extern bool initialize_display(const char* title, int w, int h);
 extern void terminate_display();
